project/finebank_backend/requirements.txt
asgiref==3.11.1
charset-normalizer==3.4.7
Django==5.2
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
pillow==12.2.0
PyJWT==2.12.1
reportlab==4.4.10
sqlparse==0.5.5
tzdata==2026.1
